fix #295: use the env var $OSTYPE in bash to make sure we install the prebuilt TinyTeX binary only for linux but not for other linux distributions such as linux-musl

+++ tools/install-bin-unix.sh
@@ -13,13 +13,17 @@ tlmgr install $TL_INSTALLED_PKGS
 OSNAME=$(uname)
 [ -z $OSNAME ] && echo "This operating system is not supported." && exit 1
 
+if [ -z $OSTYPE ]; then
+  OSTYPE=$([ -x "$(command -v bash)" ] && bash -c 'echo $OSTYPE')
+fi
+
 TINYTEX_INSTALLER=${TINYTEX_INSTALLER:-"TinyTeX-1"}
 
 if [ $OSNAME = 'Darwin' ]; then
   TEXDIR=${TINYTEX_DIR:-~/Library}/TinyTeX
 else
   TEXDIR=${TINYTEX_DIR:-~}/.TinyTeX
-  if [ $OSNAME != 'Linux' -o $(uname -m) != 'x86_64' ]; then
+  if [ $OSNAME != 'Linux' -o $(uname -m) != 'x86_64' -o "$OSTYPE" != 'linux-gnu' ]; then
     TINYTEX_INSTALLER="installer-unix"
   fi
 fi
@@ -32,19 +36,16 @@ else
   TINYTEX_URL="https://github.com/yihui/tinytex-releases/releases/download/v$TINYTEX_VERSION/$TINYTEX_INSTALLER-v$TINYTEX_VERSION"
 fi
 
-case $OSNAME in
-  "Darwin")
+if [ $OSNAME = 'Darwin']; then
     curl -L ${TINYTEX_URL}.tgz -o TinyTeX.tgz
     tar xzf TinyTeX.tgz -C $(dirname $TEXDIR)
     rm TinyTeX.tgz
-    ;;
-  "Linux")
+else if [ $TINYTEX_INSTALLER != 'installer-unix' ]; then
     wget --progress=dot:giga -O TinyTeX.tar.gz ${TINYTEX_URL}.tar.gz
     tar xzf TinyTeX.tar.gz -C $(dirname $TEXDIR)
     rm TinyTeX.tar.gz
-    ;;
-  *)
-    echo "We do not have a prebuilt TinyTeX package for the operating system ${OSNAME}."
+  else
+    echo "We do not have a prebuilt TinyTeX package for this operating system ${OSTYPE}."
     echo "I will try to install from source for you instead."
     wget ${TINYTEX_URL}.tar.gz
     tar xzf ${TINYTEX_INSTALLER}.tar.gz
@@ -52,8 +53,8 @@ case $OSNAME in
     mkdir -p $TEXDIR
     mv texlive/* $TEXDIR
     rm -r texlive ${TINYTEX_INSTALLER}.tar.gz install.sh
-    ;;
-esac
+  fi
+fi
 
 cd $TEXDIR/bin/*/
 [ $OSNAME != "Darwin" ] && ./tlmgr option sys_bin ~/bin
